requirements.txt
redis>=5.0.0
psycopg2-binary>=2.9.0
pandas>=2.2.0
pyarrow>=15.0.0
python-dotenv>=1.0.0
boto3>=1.28.0
